chore: Reorganise directory structure
