+++ symptom-checker-backend/controllers/auth/user/signup.js
@@ -1,6 +1,6 @@
 //const signupRouter = require('express').Router()
 const bcrypt = require('bcryptjs')
-const User = require('../../models/user')
+const User = require('../../../database/models/user')
 
 exports.registerUser = async(request, response) => {
     try{
