+++ symptom-checker-backend/database/models/symptom.js
