+++ symptom-checker-backend/routes/userAuthRouter.js
@@ -1,6 +1,6 @@
 const router = require('express').Router()
-const signUpController = require('../controllers/auth/signup')
-const loginController = require('../controllers/auth/login')
+const signUpController = require('../controllers/auth/user/signup')
+const loginController = require('../controllers/auth/user/login')
 
 // Signup route
 router.post('/signup', signUpController.registerUser)
